backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.7
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
django-allauth>=0.44.0
dj-rest-auth>=2.1.9
requests>=2.25.1